Move QueryPartDef out from angular app
