+++ public/app/plugins/datasource/influxdb/query_part.ts
@@ -1,5 +1,5 @@
 import { map, clone } from 'lodash';
-import { QueryPartDef, QueryPart, functionRenderer, suffixRenderer } from 'app/angular/components/query_part';
+import { QueryPartDef, QueryPart, functionRenderer, suffixRenderer } from 'app/features/alerting/state/query_part';
 
 const index: any[] = [];
 const categories: any = {
